requirements.txt
altgraph==0.17.4
macholib==1.16.3
packaging==24.0
pillow==10.3.0
pyinstaller==6.6.0
pyinstaller-hooks-contrib==2024.5
